+++ test/custom_element.test.ts
@@ -1,52 +1,22 @@
-import { defineWebComponent, html } from '../src';
-import { appendNodes, createElement } from '@fluss/web';
+import { defineWebComponent, html, WebComponent } from '../src';
 
 describe('Custom elements', () => {
   test('Defining default autonomous custom element', () => {
-    defineWebComponent('my-def-com', (rootElement) => html`<p>Hello</p>`);
+    defineWebComponent('my-def-com', class extends WebComponent {});
 
     expect(customElements.get('my-def-com')).toBeTruthy();
   });
 
-  test('Defining autonomous custom element with object description', () => {
-    defineWebComponent('my-desc-com', (rootElement) => html`<p>Hello</p>`, {
-      hooks: {
-        connected() {
-          console.log('connected');
-        },
-      },
-    });
-
-    expect(customElements.get('my-desc-com')).toBeTruthy();
-  });
-
-  test('Defining customized default custom element', () => {
-    defineWebComponent('my-p-com', [HTMLParagraphElement, 'p']);
-
-    expect(customElements.get('my-p-com')).toBeTruthy();
-  });
-
-  test('Defining customized custom element with object description', () => {
-    defineWebComponent('my-div-com', (rootElement) => html`<p>Hello</p>`, {
-      extends: {
-        constructor: HTMLDivElement,
-        tagName: 'div',
-      },
-    });
-
-    expect(customElements.get('my-div-com')).toBeTruthy();
-  });
-
-  test('Defining custom element with template as html element', () => {
+  test('Defining autonomous custom element with custom template', () => {
     defineWebComponent(
-      'my-template-com',
-      (rootElement) => html`
-        <template>
-          <span>Hello</span>
-        </template>
-      `
+      'my-desc-com',
+      class extends WebComponent {
+        template() {
+          return html`<template> <p>Hello</p></template>`;
+        }
+      }
     );
 
-    expect(customElements.get('my-template-com')).toBeTruthy();
+    expect(customElements.get('my-desc-com')).toBeTruthy();
   });
 });